add types for MapContainer component
